pyproject.toml
[project]
name = "taller-utils"
version = "0.1.0"
description = "Utilidades para procesamiento y análisis de encuestas en el Taller AHH1"
authors = [{ name = "Tu Nombre", email = "tu.email@ejemplo.com" }]
requires-python = ">=3.8"

dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "scipy>=1.7.0",
    "pyyaml>=5.4",
    "scikit-learn>=0.24.0"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
